Backened/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
asyncpg==0.30.0
attrs==25.3.0
bcrypt==5.0.0
black==25.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
coverage==7.10.7
cryptography==46.0.1
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.3.1
djangorestframework==3.16.1
email-validator==2.2.0
django-filter==24.3
ecdsa==0.19.1
fastapi==0.117.1
flake8==7.3.0
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-experimental==0.3.4
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langsmith==0.4.31
langdetect==1.0.9
markdown-it-py==4.0.0
marshmallow==3.26.1
mccabe==0.7.0
mdurl==0.1.2
multidict==6.6.4
mypy==1.18.2
mypy_extensions==1.1.0
npm==0.1.1
numpy==2.3.3
openai==1.109.1
optional-django==0.1.0
orjson==3.11.3
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
postgrest==2.20.0
prometheus_client==0.23.1
propcache==0.3.2
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.1.10
PyYAML==6.0.3
realtime==2.20.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
sentry-sdk==2.39.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.48.0
storage3==2.20.0
StrEnum==0.4.15
structlog==25.4.0
supabase==2.20.0
supabase-auth==2.20.0
supabase-functions==2.20.0
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.25.0
pylint==3.3.4
